refactor: split api into independent files
