+++ pages/about/index.tsx
@@ -1,6 +1,7 @@
 import About from '../../components/templates/About';
 import PageMeta from '../../components/atoms/PageMeta';
-import { getGraduates, getSeos } from '../../lib/api';
+import { getSeos } from 'api/seos';
+import { getGraduates } from 'api/graduates';
 import { CURRENT_SESSION } from '../../util/common/setting';
 
 export async function getServerSideProps() {
