+++ pages/faq/index.tsx
@@ -1,6 +1,7 @@
 import Faq from 'components/templates/Faq';
 import PageMeta from 'components/atoms/PageMeta';
-import { getSeos, getFaqs } from 'lib/api';
+import { getFaqs } from 'api/faqs';
+import { getSeos } from 'api/seos';
 
 export async function getServerSideProps() {
   const [meta, faqs] = await Promise.all([getSeos('faq'), getFaqs()]);
