+++ pages/index.tsx
@@ -1,6 +1,6 @@
 import Home from './home';
 import PageMeta from '../components/atoms/PageMeta';
-import { getSeos } from '../lib/api';
+import { getSeos } from 'api/seos';
 import { Meta } from '../components/atoms/PageMeta/PageMeta';
 
 export async function getServerSideProps() {
